+++ jtl/outproc.go
@@ -34,10 +34,13 @@ func handleEvent(ctx Context, stats *ServiceStats, event *JDoc, raw string, debu
 	}
 	initialCtx := ctx
 	ctx = ctx.SubContext()
+	allowPartner := GetConfig(ctx).AllowPartner
+	defaultPartner := GetConfig(ctx).DefaultPartner
 	var wg sync.WaitGroup
 	for _, handler := range handlers {
 		ctx.AddLogValue("topic", handler.Topic)
-		ctx.AddLogValue(LogTenantId, handler.TenantId)
+		ctx.AddLogValue(LogTenantId, ExtractAppId(handler.TenantId, allowPartner))
+		ctx.AddLogValue(LogPartnerId, ExtractPartnerId(handler.TenantId, allowPartner, defaultPartner))
 		ctx.AddLogValue("handler", handler.Name)
 		publishers, err := handler.ProcessEvent(initialCtx.SubContext(), event)
 		if err != nil {