+++ CHANGELOG.md
@@ -4,11 +4,31 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This `CHANGELOG.md` implements the spirit of http://keepachangelog.com/.
 
-## [1.40](https://github.com/Comcast/eel/compare/v1.39.0...dev) - [Unreleased]
+## [1.42](https://github.com/Comcast/eel/compare/v1.41.0...dev) - [Unreleased]
+
+## [1.41](https://github.com/Comcast/eel/compare/v1.40.0...v1.41.0) - 2021-06-02
+
+### Added
+* Logging for partner Id
+* Covert to go mod
+
+## [1.40](https://github.com/Comcast/eel/compare/v1.39.0...v1.40.0) - 2021-05-10
+
+### Added
+* AllowPartner config parameter
+* partner() function
+* Add logic to handle http partner header
 
 ## [1.39](https://github.com/Comcast/eel/compare/v1.38.0...v1.39.0) - 2020-12-04
 
 ## [1.38](https://github.com/Comcast/eel/compare/v1.37.0...v1.38.0) - 2020-11-16
+
+### Added
+* XRULES-20200606: Add function 'param' to read query string parameters of incoming event url
+* XRULES-20200606: Add use of HTTP_PROXY environment variable to send data via a proxy
+
+### Updated
+* XRULES-20200606: Docker build; Make smaller runtime image by using multi-stage build, and enable testing
 
 ## [1.37](https://github.com/Comcast/eel/compare/v1.36.0...v1.37.0) - 2020-03-09